move base functionality to NetManager class + cleanup
