+++ src/JoinResponsePacket.cpp
@@ -1,4 +1,4 @@
-#include "JoinResponsePacket.h"
+#include "Main.h"
 
 void JoinResponsePacket::setResponse(JoinResponse joinResponse){
