+++ src/Main.cpp
@@ -1,0 +1,12 @@
+#include "Main.h"
+
+int main(int argc, char* argv[])
+{
+        
+    NetManager netManager;
+    netManager.update();
+
+    
+    return 0;
+}
+
